refactor: types, common 폴더 아래로 이동
