+++ src/workspaces/controllers/workspaces.controller.ts
@@ -16,7 +16,7 @@ import { ApiBearerAuth, ApiOperation, ApiParam, ApiResponse, ApiTags } from '@ne
 import { WorkspaceRoles } from '../../auth/decorators/workspace-role.decorator';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
 import { WorkspaceRoleGuard } from '../../auth/guards/workspace-role.guard';
-import { AuthenticatedRequest } from '../../types/authenticated-request';
+import { AuthenticatedRequest } from '../../common/types/authenticated-request';
 import { AddUserToWorkspaceDto } from '../dto/add-user-to-workspace.dto';
 import { CreateWorkspaceDto } from '../dto/create-workspace.dto';
 import { UpdateWorkspaceDto } from '../dto/update-workspace.dto';
@@ -57,7 +57,7 @@ export class WorkspacesController {
 	@ApiResponse({ status: 400, description: '잘못된 요청입니다.' })
 	async create(
 		@Body() createWorkspaceDto: CreateWorkspaceDto,
-		@Request() req: AuthenticatedRequest, // 👈 타입 명시
+		@Request() req: AuthenticatedRequest,
 	): Promise<WorkspaceCreateResponseDto> {
 		return this.workspacesService.create(createWorkspaceDto, req.user.sub);
 	}
